git-svn-id: https://swt-chart.svn.sourceforge.net/svnroot/swt-chart/trunk@252 d30752db-1043-0410-9340-8085d4ee3fb9

+++ org.swtchart/src/org/swtchart/internal/axis/AxisTickLabels.java
@@ -874,7 +874,7 @@ public class AxisTickLabels implements PaintListener {
      *            the angle
      */
     private void drawRotatedText(GC gc, String text, float x, float y,
-            float angle) {
+            int angle) {
 
         int textWidth = gc.textExtent(text).x;
         int textHeight = gc.textExtent(text).y;
